Login & Sign Up Support
Began Mongo database API. Sign up page is successfully creating new record in the users collection in db. Login is successfully verifying username and encrypted password and logging user in to the main home page. When a user is not logged in, only the Login and Signup buttons will show up in the navbar. When a user logs in, they will only be shown the Home and Logout buttons in the navbar.