Using flask_login
We are now using flask_login to manage logging in to a users account. Users are uniquely identified by their username. REMEMBER: mongodb does NOT work on FSU wifi, so proper testing has to be done at home or through a VPN. Sign up and login are working properly, password is not hidden at this time.