Pulling all card data from forms
now we are pulling all the cards into a dictionary where the term is the key and the definition is the value. We can just pass that to a function (not existent yet) in the database file that will take the dict of cards, title of cardset, and username of user to store in mongo db.