Admin/Mod/User Editing & Deleting Sets
Admin, moderators, users now see the respective card sets in the viewing window when choosing either to view or edit a card set. Also added a delete set button on the edit page BUUUTTT I could not get that trash can to be on the LEFT side of the set button, so it is on top and its terrible.