Account Management Start
Made a page for account management, only viewable when logged in to admin account. Has initial table and some simple css to give it the general look. The delete button I would like in the future to be stevens modal trash button functionality but for now is a delete button. The button to the right will be the username of the user account and when clicked it will allow editing of the user information. Notice the delete button is only by the user as of now, maybe moderators should get a delete button too.