Delete Acc/ Edit Priv
User functionality for being able to edit a users priviledge, and delete a users account has been implemented. The functionality to TRULY delete it from the DB is not implemented. Just need to make 2 funcs in db file to delete an acc given a name, and to change privs given name and new priv status.